util として分離する
